Moved ping management into seperate file
